changed track length column to a double. Worked on seeder, but need to work on Factories for teams and tracks.

+++ database/migrations/2023_04_13_045611_create_tracks_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('location');
-            $table->decimal('trackLength');
+            $table->double('trackLength', 4, 2);
             $table->timestamps();
         });
     }
